+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased
 
 * Bumped setuptools to 70.3.0 to fix CVE-2024-6345.
 
+* Fixed Grand Central initContainer image update.
+
 2.40.0 (2024-07-17)
 -------------------